Move the view session recording test mocks to a separate file
